chat: Split off ChatInput in its own container.
